masterpig-scanner-gitready/requirements.txt
python-dotenv==1.0.1
pymongo==4.7.3
motor==3.6.0
bip-utils==2.9.3
aiohttp==3.10.5
fastapi==0.112.2
uvicorn==0.30.6
orjson==3.10.7
pydantic==2.8.2

jinja2==3.1.4
